+++ app/Jobs/DownloadLessonAudioJob.php
@@ -80,6 +80,7 @@ class DownloadLessonAudioJob implements ShouldBeUnique, ShouldQueue
             $pipelineRunService->dispatchQueuedRuns($lesson);
         } catch (Throwable $exception) {
             $lesson = $this->markAsFailed($lesson, $exception->getMessage());
+            $pipelineRunService->setDownloadErrorOnFirstStep($lesson, $exception->getMessage());
             $eventBroadcaster->downloadFailed($lesson, $exception->getMessage());
 
             throw $exception;